requirements.txt
requests==2.26.0
pytest==6.2.5
cerberus==1.3.4
allure-pytest==2.9.44